requirements.txt
langgraph>=1.0.2,<1.1.0
langchain-anthropic
langchain-core
boto3==1.35.0
python-dotenv==1.0.1
rich==13.7.1
reportlab==4.2.2
slack-sdk>=3.38.0
pandas==2.2.3
numpy>=2.0.0,<2.3.0
